Fixed bugs, organised tests in folders accordingly
