+++ tests/Http/Controllers/Api/Entries/Delete/DeleteControllerEntryLocalTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Http\Controllers\Api\Entries;
+namespace Tests\Http\Controllers\Api\Entries\Delete;
 
 use ec5\Libraries\Generators\EntryGenerator;
 use ec5\Libraries\Generators\ProjectDefinitionGenerator;
@@ -18,7 +18,7 @@ use Ramsey\Uuid\Uuid;
 use Tests\TestCase;
 use Throwable;
 
-class DeleteControllerEntryS3Test extends TestCase
+class DeleteControllerEntryLocalTest extends TestCase
 {
     use DatabaseTransactions;
 
@@ -95,8 +95,8 @@ class DeleteControllerEntryS3Test extends TestCase
             $this->projectDefinition = $projectDefinition;
             $this->entryGenerator = $entryGenerator;
 
-            //set storage (and all disks) to S3
-            $this->overrideStorageDriver('s3');
+            //set storage (and all disks) to local storage
+            $this->overrideStorageDriver('local');
         } catch (Throwable $e) {
             $this->logTestError($e, $response);
         }
@@ -171,16 +171,10 @@ class DeleteControllerEntryS3Test extends TestCase
         );
 
         //add a fake file per each entry (per each media type)
-        //photo
-        Storage::disk('photo')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.jpg', str_repeat('A', 1234));
-        //audio
-        Storage::disk('audio')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 1234));
-        //video
-        Storage::disk('video')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 1234));
+        $this->addMediaFilesToEntry($this->project->ref, $entry->uuid);
 
         $projectStats = ProjectStats::where('project_id', $this->project->id)->firstOrFail();
         $projectStats->updateEntryCounters($this->project->id);
-
         // collect stats
         $photoFiles = Storage::disk('photo')->files($this->project->ref);
         $audioFiles = Storage::disk('audio')->files($this->project->ref);
@@ -200,16 +194,18 @@ class DeleteControllerEntryS3Test extends TestCase
             sizeof($videoFiles)
         );
 
+        //assert project stats are correct after adding media files
         $updatedProjectStats = ProjectStats::where('project_id', $this->project->id)->first();
         $totalBytes = $photoBytes + $audioBytes + $videoBytes;
+        $totalFiles = sizeof($photoFiles) + sizeof($audioFiles) + sizeof($videoFiles);
         $this->assertEquals($totalBytes, $updatedProjectStats->total_bytes);
-        $this->assertEquals(3, $updatedProjectStats->total_files);
+        $this->assertEquals($totalFiles, $updatedProjectStats->total_files);
         $this->assertEquals($photoBytes, $updatedProjectStats->photo_bytes);
-        $this->assertEquals(count($photoFiles), $updatedProjectStats->photo_files);
+        $this->assertEquals(sizeof($photoFiles), $updatedProjectStats->photo_files);
         $this->assertEquals($audioBytes, $updatedProjectStats->audio_bytes);
-        $this->assertEquals(count($audioFiles), $updatedProjectStats->audio_files);
+        $this->assertEquals(sizeof($audioFiles), $updatedProjectStats->audio_files);
         $this->assertEquals($videoBytes, $updatedProjectStats->video_bytes);
-        $this->assertEquals(count($videoFiles), $updatedProjectStats->video_files);
+        $this->assertEquals(sizeof($videoFiles), $updatedProjectStats->video_files);
 
         $payload = $this->createPayload(
             $this->user->id,
@@ -224,7 +220,6 @@ class DeleteControllerEntryS3Test extends TestCase
             ProjectStats::where('project_id', $this->project->id)
                 ->value('total_entries')
         );
-
 
         //hit the delete endpoint
         $response = [];
@@ -258,18 +253,16 @@ class DeleteControllerEntryS3Test extends TestCase
             $videos = Storage::disk('video')->files($this->project->ref);
             $this->assertCount(0, $videos);
 
-            //assert media storage stats are updated
+            //assert media stats are zeroed
             $updatedProjectStats = ProjectStats::where('project_id', $this->project->id)->first();
-            $this->assertEquals(0, (int) $updatedProjectStats->total_bytes);
-            $this->assertEquals(0, (int) $updatedProjectStats->total_files);
-            $this->assertEquals(0, (int) $updatedProjectStats->total_entries);
-            $this->assertEquals(0, (int) $updatedProjectStats->photo_bytes);
-            $this->assertEquals(0, (int) $updatedProjectStats->photo_files);
-            $this->assertEquals(0, (int) $updatedProjectStats->audio_bytes);
-            $this->assertEquals(0, (int) $updatedProjectStats->audio_files);
-            $this->assertEquals(0, (int) $updatedProjectStats->video_bytes);
-            $this->assertEquals(0, (int) $updatedProjectStats->video_files);
-
+            $this->assertEquals(0, $updatedProjectStats->total_bytes);
+            $this->assertEquals(0, $updatedProjectStats->total_files);
+            $this->assertEquals(0, $updatedProjectStats->photo_bytes);
+            $this->assertEquals(0, $updatedProjectStats->photo_files);
+            $this->assertEquals(0, $updatedProjectStats->audio_bytes);
+            $this->assertEquals(0, $updatedProjectStats->audio_files);
+            $this->assertEquals(0, $updatedProjectStats->video_bytes);
+            $this->assertEquals(0, $updatedProjectStats->video_files);
 
 
         } catch (Throwable $e) {
@@ -358,13 +351,8 @@ class DeleteControllerEntryS3Test extends TestCase
             ]
         );
 
-        //add a fake file per each parent entry (per each media type)
-        //photo
-        Storage::disk('photo')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.jpg', str_repeat('A', 1234));
-        //audio
-        Storage::disk('audio')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 1234));
-        //video
-        Storage::disk('video')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 1234));
+        //add a fake file per each entry (per each media type)
+        $this->addMediaFilesToEntry($this->project->ref, $entry->uuid);
 
         $childFormRef = $this->projectDefinition['data']['project']['forms'][1]['ref'];
         $childEntry = factory(Entry::class)->create(
@@ -379,13 +367,12 @@ class DeleteControllerEntryS3Test extends TestCase
             ]
         );
 
-        //add a fake file per each child entry (per each media type)
-        //photo
-        Storage::disk('photo')->put($this->project->ref . '/' . $childEntry->uuid . '_' . time() . '.jpg', str_repeat('A', 1234));
-        //audio
-        Storage::disk('audio')->put($this->project->ref . '/' . $childEntry->uuid . '_' . time() . '.mp4', str_repeat('A', 1234));
-        //video
-        Storage::disk('video')->put($this->project->ref . '/' . $childEntry->uuid . '_' . time() . '.mp4', str_repeat('A', 1234));
+        //add a fake file per each entry (per each media type)
+        $this->addMediaFilesToEntry($this->project->ref, $childEntry->uuid);
+
+        //update $entry with child count
+        $entry->child_counts = 1;
+        $entry->save();
 
         $projectStats = ProjectStats::where('project_id', $this->project->id)->firstOrFail();
         $projectStats->updateEntryCounters($this->project->id);
@@ -409,22 +396,18 @@ class DeleteControllerEntryS3Test extends TestCase
             sizeof($videoFiles)
         );
 
+        //assert project stats are correct after adding media files
         $updatedProjectStats = ProjectStats::where('project_id', $this->project->id)->first();
         $totalBytes = $photoBytes + $audioBytes + $videoBytes;
-        $totalFiles = sizeof($photoFiles) + sizeOf($videoFiles) + sizeof($audioFiles);
+        $totalFiles = sizeof($photoFiles) + sizeof($audioFiles) + sizeof($videoFiles);
         $this->assertEquals($totalBytes, $updatedProjectStats->total_bytes);
         $this->assertEquals($totalFiles, $updatedProjectStats->total_files);
         $this->assertEquals($photoBytes, $updatedProjectStats->photo_bytes);
-        $this->assertEquals(count($photoFiles), $updatedProjectStats->photo_files);
+        $this->assertEquals(sizeof($photoFiles), $updatedProjectStats->photo_files);
         $this->assertEquals($audioBytes, $updatedProjectStats->audio_bytes);
-        $this->assertEquals(count($audioFiles), $updatedProjectStats->audio_files);
+        $this->assertEquals(sizeof($audioFiles), $updatedProjectStats->audio_files);
         $this->assertEquals($videoBytes, $updatedProjectStats->video_bytes);
-        $this->assertEquals(count($videoFiles), $updatedProjectStats->video_files);
-
-
-        //update $entry with child count
-        $entry->child_counts = 1;
-        $entry->save();
+        $this->assertEquals(sizeof($videoFiles), $updatedProjectStats->video_files);
 
         $this->assertCount(1, Entry::where('uuid', $entry->uuid)->get());
         $this->assertEquals(1, Entry::where('uuid', $entry->uuid)->value('child_counts'));
@@ -489,22 +472,32 @@ class DeleteControllerEntryS3Test extends TestCase
             $videos = Storage::disk('video')->files($this->project->ref);
             $this->assertCount(1, $videos);
 
+            // recompute expected remaining sizes and files from S3
+            $photoBytesRemaining = collect($photos)->map(fn ($f) => Storage::disk('photo')->size($f))->sum();
+            $audioBytesRemaining = collect($audios)->map(fn ($f) => Storage::disk('audio')->size($f))->sum();
+            $videoBytesRemaining = collect($videos)->map(fn ($f) => Storage::disk('video')->size($f))->sum();
+
+            $totalBytesRemaining = $photoBytesRemaining + $audioBytesRemaining + $videoBytesRemaining;
+            $totalFilesRemaining = count($photos) + count($audios) + count($videos);
+
+            //Assert ProjectStats matches remaining media
+            $stats = ProjectStats::where('project_id', $this->project->id)->first();
+            $this->assertEquals($photoBytesRemaining, (int) $stats->photo_bytes);
+            $this->assertEquals(count($photos), (int) $stats->photo_files);
+
+            $this->assertEquals($audioBytesRemaining, (int) $stats->audio_bytes);
+            $this->assertEquals(count($audios), (int) $stats->audio_files);
+
+            $this->assertEquals($videoBytesRemaining, (int) $stats->video_bytes);
+            $this->assertEquals(count($videos), (int) $stats->video_files);
+
+            $this->assertEquals($totalBytesRemaining, (int) $stats->total_bytes);
+            $this->assertEquals($totalFilesRemaining, (int) $stats->total_files);
+
             //delete fake files
             Storage::disk('photo')->deleteDirectory($this->project->ref);
             Storage::disk('audio')->deleteDirectory($this->project->ref);
             Storage::disk('video')->deleteDirectory($this->project->ref);
-
-            //assert media storage stats are updated
-            $updatedProjectStats = ProjectStats::where('project_id', $this->project->id)->first();
-            $this->assertEquals(3 * 1234, (int) $updatedProjectStats->total_bytes);
-            $this->assertEquals(3, (int) $updatedProjectStats->total_files);
-            $this->assertEquals(1, (int) $updatedProjectStats->total_entries);
-            $this->assertEquals(1234, (int) $updatedProjectStats->photo_bytes);
-            $this->assertEquals(1, (int) $updatedProjectStats->photo_files);
-            $this->assertEquals(1234, (int) $updatedProjectStats->audio_bytes);
-            $this->assertEquals(1, (int) $updatedProjectStats->audio_files);
-            $this->assertEquals(1234, (int) $updatedProjectStats->video_bytes);
-            $this->assertEquals(1, (int) $updatedProjectStats->video_files);
 
         } catch (Throwable $e) {
             $this->logTestError($e, $response);
@@ -526,17 +519,11 @@ class DeleteControllerEntryS3Test extends TestCase
         );
 
         //add a fake file per each entry (per each media type)
-        //photo
-        Storage::disk('photo')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.jpg', str_repeat('A', 1024));
-        //audio
-        Storage::disk('audio')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 2048));
-        //video
-        Storage::disk('video')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 4096));
+        $this->addMediaFilesToEntry($this->project->ref, $entry->uuid);
 
         $uuids[] = $entry->uuid;
 
         $parentEntry = $entry; // Initial parent entry
-        //add child entries ($i starts from 1 as 0 is the parent entry)
         for ($i = 1; $i < config('epicollect.limits.formsMaxCount'); $i++) {
             $childFormRef = $forms[$i]['ref'];
             $childEntry = factory(Entry::class)->create([
@@ -550,12 +537,7 @@ class DeleteControllerEntryS3Test extends TestCase
             ]);
 
             //add a fake file per each entry (per each media type)
-            //photo
-            Storage::disk('photo')->put($this->project->ref . '/' . $childEntry->uuid . '_' . time() . '.jpg', str_repeat('A', 1024));
-            //audio
-            Storage::disk('audio')->put($this->project->ref . '/' . $childEntry->uuid . '_' . time() . '.mp4', str_repeat('A', 2048));
-            //video
-            Storage::disk('video')->put($this->project->ref . '/' . $childEntry->uuid . '_' . time() . '.mp4', str_repeat('A', 4096));
+            $this->addMediaFilesToEntry($this->project->ref, $childEntry->uuid);
 
             $uuids[] = $childEntry->uuid;
 
@@ -563,39 +545,6 @@ class DeleteControllerEntryS3Test extends TestCase
             // Set the created child entry as the parent for the next iteration
             $parentEntry = $childEntry;
         }
-
-        $projectStats = ProjectStats::where('project_id', $this->project->id)->firstOrFail();
-        $projectStats->updateEntryCounters($this->project->id);
-
-        // collect stats
-        $photoFiles = Storage::disk('photo')->files($this->project->ref);
-        $audioFiles = Storage::disk('audio')->files($this->project->ref);
-        $videoFiles = Storage::disk('video')->files($this->project->ref);
-
-        $photoBytes = collect($photoFiles)->map(fn ($f) => Storage::disk('photo')->size($f))->sum();
-        $audioBytes = collect($audioFiles)->map(fn ($f) => Storage::disk('audio')->size($f))->sum();
-        $videoBytes = collect($videoFiles)->map(fn ($f) => Storage::disk('video')->size($f))->sum();
-
-        //update media storage stats
-        $projectStats->incrementMediaStorageUsage(
-            $photoBytes,
-            sizeof($photoFiles),
-            $audioBytes,
-            sizeof($audioFiles),
-            $videoBytes,
-            sizeof($videoFiles)
-        );
-
-        $updatedProjectStats = ProjectStats::where('project_id', $this->project->id)->first();
-        $totalBytes = $photoBytes + $audioBytes + $videoBytes;
-        $this->assertEquals($totalBytes, $updatedProjectStats->total_bytes);
-        $this->assertEquals(3 * config('epicollect.limits.formsMaxCount'), $updatedProjectStats->total_files);
-        $this->assertEquals($photoBytes, $updatedProjectStats->photo_bytes);
-        $this->assertEquals(count($photoFiles), $updatedProjectStats->photo_files);
-        $this->assertEquals($audioBytes, $updatedProjectStats->audio_bytes);
-        $this->assertEquals(count($audioFiles), $updatedProjectStats->audio_files);
-        $this->assertEquals($videoBytes, $updatedProjectStats->video_bytes);
-        $this->assertEquals(count($videoFiles), $updatedProjectStats->video_files);
 
 
         //update $entry with child count
@@ -683,17 +632,16 @@ class DeleteControllerEntryS3Test extends TestCase
             $videos = Storage::disk('video')->files($this->project->ref);
             $this->assertCount(0, $videos);
 
-            //assert media storage stats are updated
+            //assert media stats are zeroed
             $updatedProjectStats = ProjectStats::where('project_id', $this->project->id)->first();
-            $this->assertEquals(0, (int) $updatedProjectStats->total_bytes);
-            $this->assertEquals(0, (int) $updatedProjectStats->total_files);
-            $this->assertEquals(0, (int) $updatedProjectStats->total_entries);
-            $this->assertEquals(0, (int) $updatedProjectStats->photo_bytes);
-            $this->assertEquals(0, (int) $updatedProjectStats->photo_files);
-            $this->assertEquals(0, (int) $updatedProjectStats->audio_bytes);
-            $this->assertEquals(0, (int) $updatedProjectStats->audio_files);
-            $this->assertEquals(0, (int) $updatedProjectStats->video_bytes);
-            $this->assertEquals(0, (int) $updatedProjectStats->video_files);
+            $this->assertEquals(0, $updatedProjectStats->total_bytes);
+            $this->assertEquals(0, $updatedProjectStats->total_files);
+            $this->assertEquals(0, $updatedProjectStats->photo_bytes);
+            $this->assertEquals(0, $updatedProjectStats->photo_files);
+            $this->assertEquals(0, $updatedProjectStats->audio_bytes);
+            $this->assertEquals(0, $updatedProjectStats->audio_files);
+            $this->assertEquals(0, $updatedProjectStats->video_bytes);
+            $this->assertEquals(0, $updatedProjectStats->video_files);
 
         } catch (Throwable $e) {
             $this->logTestError($e, $response);
@@ -715,25 +663,15 @@ class DeleteControllerEntryS3Test extends TestCase
         );
 
         //add a fake file per each entry (per each media type)
-        //photo
-        Storage::disk('photo')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.jpg', str_repeat('A', 1024));
-        //audio
-        Storage::disk('audio')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 2048));
-        //video
-        Storage::disk('video')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 4096));
+        $this->addMediaFilesToEntry($this->project->ref, $entry->uuid);
 
-        //add a few media files to test they will not bwe deleted
+        //add a few media files to test they will not be deleted
         $numOfFilesNotToBeDelete = rand(3, 10);
         for ($j = 0; $j < $numOfFilesNotToBeDelete; $j++) {
             $uuid = Uuid::uuid4()->toString();
-            //photo
-            Storage::disk('photo')->put($this->project->ref . '/' . $uuid . '_' . time() . '.jpg', str_repeat('A', 1024));
-            //audio
-            Storage::disk('audio')->put($this->project->ref . '/' . $uuid . '_' . time() . '.mp4', str_repeat('A', 2048));
-            //video
-            Storage::disk('video')->put($this->project->ref . '/' . $uuid . '_' . time() . '.mp4', str_repeat('A', 4096));
+            //add a fake file per each entry (per each media type)
+            $this->addMediaFilesToEntry($this->project->ref, $uuid);
         }
-
 
         $inputs = $forms[0]['inputs'];
         $branchRef = '';
@@ -761,22 +699,8 @@ class DeleteControllerEntryS3Test extends TestCase
             $branchUuids[] = $branchEntry->uuid;
 
             //add a fake file per each entry (per each media type)
-            //photo
-            Storage::disk('photo')->put($this->project->ref . '/' . $branchEntry->uuid . '_' . time() . '.jpg', str_repeat('A', 1234));
-            //audio
-            Storage::disk('audio')->put($this->project->ref . '/' . $branchEntry->uuid . '_' . time() . '.mp4', str_repeat('A', 2048));
-            //video
-            Storage::disk('video')->put($this->project->ref . '/' . $branchEntry->uuid . '_' . time() . '.mp4', str_repeat('A', 4096));
+            $this->addMediaFilesToEntry($this->project->ref, $branchEntry->uuid);
         }
-
-        $branchCounts = [
-            $branchRef => $numOfBranchEntries
-        ];
-        $entry->branch_counts = json_encode($branchCounts);
-        $entry->save();
-
-        $this->assertCount(1, Entry::where('project_id', $this->project->id)->get());
-        $this->assertCount($numOfBranchEntries, BranchEntry::where('project_id', $this->project->id)->get());
 
         $projectStats = ProjectStats::where('project_id', $this->project->id)->firstOrFail();
 
@@ -811,6 +735,15 @@ class DeleteControllerEntryS3Test extends TestCase
         $this->assertEquals(sizeof($audioFiles), $updatedProjectStats->audio_files);
         $this->assertEquals($videoBytes, $updatedProjectStats->video_bytes);
         $this->assertEquals(sizeof($videoFiles), $updatedProjectStats->video_files);
+
+        $branchCounts = [
+            $branchRef => $numOfBranchEntries
+        ];
+        $entry->branch_counts = json_encode($branchCounts);
+        $entry->save();
+
+        $this->assertCount(1, Entry::where('project_id', $this->project->id)->get());
+        $this->assertCount($numOfBranchEntries, BranchEntry::where('project_id', $this->project->id)->get());
 
 
         //build delete payload
@@ -877,8 +810,6 @@ class DeleteControllerEntryS3Test extends TestCase
             Storage::disk('photo')->deleteDirectory($this->project->ref);
             Storage::disk('audio')->deleteDirectory($this->project->ref);
             Storage::disk('video')->deleteDirectory($this->project->ref);
-
-
         } catch (Throwable $e) {
             $this->logTestError($e, $response);
         }
@@ -898,15 +829,7 @@ class DeleteControllerEntryS3Test extends TestCase
         );
 
         //add a fake file per each entry (per each media type)
-        //photo
-        Storage::disk('photo')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.jpg', str_repeat('A', 1024));
-        //audio
-        Storage::disk('audio')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 2048));
-        //video
-        Storage::disk('video')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 4096));
-
-        $projectStats = ProjectStats::where('project_id', $this->project->id)->firstOrFail();
-        $projectStats->updateEntryCounters($this->project->id);
+        $this->addMediaFilesToEntry($this->project->ref, $entry->uuid);
 
         //add branch refs
         $branchRefs = [];
@@ -934,20 +857,16 @@ class DeleteControllerEntryS3Test extends TestCase
                     'uuid' => $branchUuid
                 ]
             );
-            //photo
-            Storage::disk('photo')->put($this->project->ref . '/' . $branchUuid . '_' . time() . '.jpg', str_repeat('A', 1024));
-            //audio
-            Storage::disk('audio')->put($this->project->ref . '/' . $branchUuid . '_' . time() . '.mp4', str_repeat('A', 1024));
-            //video
-            Storage::disk('video')->put($this->project->ref . '/' . $branchUuid . '_' . time() . '.mp4', str_repeat('A', 1024));
+            //add a fake file per each entry (per each media type)
+            $this->addMediaFilesToEntry($this->project->ref, $branchUuid);
 
             $branchCounts[$branchRef] = 1;
         }
         $entry->branch_counts = json_encode($branchCounts);
         $entry->save();
+
         $this->assertCount(1, Entry::where('project_id', $this->project->id)->get());
         $this->assertCount(sizeof($branchEntries), BranchEntry::where('project_id', $this->project->id)->get());
-
 
         $projectStats = ProjectStats::where('project_id', $this->project->id)->firstOrFail();
 
@@ -1031,7 +950,6 @@ class DeleteControllerEntryS3Test extends TestCase
             $videosRemaining = Storage::disk('video')->files($this->project->ref);
             $this->assertCount(1 + (sizeof($branchEntries) - 1), $videosRemaining);
 
-
             // recompute expected remaining sizes and files from S3
             $photoBytesRemaining = collect($photosRemaining)->map(fn ($f) => Storage::disk('photo')->size($f))->sum();
             $audioBytesRemaining = collect($audiosRemaining)->map(fn ($f) => Storage::disk('audio')->size($f))->sum();
@@ -1079,12 +997,7 @@ class DeleteControllerEntryS3Test extends TestCase
         );
 
         //add a fake file per each entry (per each media type)
-        //photo
-        Storage::disk('photo')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.jpg', str_repeat('A', 1024));
-        //audio
-        Storage::disk('audio')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 2048));
-        //video
-        Storage::disk('video')->put($this->project->ref . '/' . $entry->uuid . '_' . time() . '.mp4', str_repeat('A', 4096));
+        $this->addMediaFilesToEntry($this->project->ref, $entry->uuid);
 
         //add branch refs (only child form)
         $branchRefs = [];
@@ -1114,12 +1027,7 @@ class DeleteControllerEntryS3Test extends TestCase
                     'uuid' => $branchUuid
                 ]
             );
-            //photo
-            Storage::disk('photo')->put($this->project->ref . '/' . $branchUuid . '_' . time() . '.jpg', str_repeat('A', 1234));
-            //audio
-            Storage::disk('audio')->put($this->project->ref . '/' . $branchUuid . '_' . time() . '.mp4', str_repeat('A', 1234));
-            //video
-            Storage::disk('video')->put($this->project->ref . '/' . $branchUuid . '_' . time() . '.mp4', str_repeat('A', 1234));
+            $this->addMediaFilesToEntry($this->project->ref, $branchUuid);
 
             $branchCounts[$branchRef] = 1;
         }
@@ -1182,21 +1090,11 @@ class DeleteControllerEntryS3Test extends TestCase
             Storage::disk('audio')->deleteDirectory($this->project->ref);
             Storage::disk('video')->deleteDirectory($this->project->ref);
 
-            //assert project stats are updated after deleting entry and branch entries
-            $updatedProjectStats = ProjectStats::where('project_id', $this->project->id)->first();
-            $this->assertEquals(0, $updatedProjectStats->total_bytes);
-            $this->assertEquals(0, $updatedProjectStats->total_files);
-            $this->assertEquals(0, $updatedProjectStats->photo_bytes);
-            $this->assertEquals(0, $updatedProjectStats->photo_files);
-            $this->assertEquals(0, $updatedProjectStats->audio_bytes);
-            $this->assertEquals(0, $updatedProjectStats->audio_files);
-            $this->assertEquals(0, $updatedProjectStats->video_bytes);
-            $this->assertEquals(0, $updatedProjectStats->video_files);
-
         } catch (Throwable $e) {
             $this->logTestError($e, $response);
         }
     }
+
 
     private function createPayload($userId, $entryUuid, $formRef, $branchCounts, $childCounts)
     {
@@ -1224,5 +1122,15 @@ class DeleteControllerEntryS3Test extends TestCase
                 'entry_uuid' => $entryUuid
             ]
         ];
+    }
+
+    private function addMediaFilesToEntry(string $projectRef, string $uuid): void
+    {
+        //photo
+        Storage::disk('photo')->put($projectRef . '/' . $uuid . '_' . time() . '.jpg', str_repeat('A', 1024));
+        //audio
+        Storage::disk('audio')->put($projectRef . '/' . $uuid . '_' . time() . '.mp4', str_repeat('A', 2048));
+        //video
+        Storage::disk('video')->put($projectRef . '/' . $uuid . '_' . time() . '.mp4', str_repeat('A', 4096));
     }
 }
